+++ .env.example
@@ -23,7 +23,7 @@ ANALYTICS_URL=""
 # ------------
 TICKETS_GOOGLE_SHEET_ID=""
 # Relative to google.js file
-TICKETS_KEY_PATH="../google-api-key.json"
+TICKETS_KEY_PATH="../../google-api-key.json"
 
 
 # Auth0 settings
